requirements.txt
pyserial
pycaw; sys_platform == 'win32'
